+++ api/models/device.py
@@ -1,24 +1,27 @@
+from flask import Flask
+from marshmallow import Schema, fields, pre_load, validate
+from flask_marshmallow import Marshmallow
 from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from app import db
+
+ma = Marshmallow()
+db = SQLAlchemy()
 
 class Device(db.Model):
+     __tablename__ = 'outlet_devices'
      id = db.Column(db.Integer, primary_key=True)
-     uuid = db.Column(db.Integer)
-     name = db.Column(db.String(32))
-     on_value = db.Column(db.Integer)
-     off_value = db.Column(db.Integer)
-     type = db.Column(db.String(16))
+     uuid = db.Column(db.Integer, unique=True)
+     name = db.Column(db.String(128), nullable=False)
+     type = db.Column(db.Integer, nullable=False)
 
-     def __init__(self, uuid, name, on_value, off_value, type):
-        self.uuid = uuid
-        self.name = name
-        self.on_value = on_value
-        self.off_value = off_value
-        self.type = type
+     def __init__(self, name, type, uuid):
+         self.name = name
+         self.type = type
+         self.uuid = uuid
 
-     def serialize(self):
-        return {"id": self.id, "name": self.name}
+class DeviceSchema(ma.Schema):
+     id = fields.Integer(dump_only=True)
+     uuid = fields.Integer(unique=True)
+     name = fields.String(required=True, validate = validate.Length(1))
+     type = fields.Integer(required=True)
 
-     def __repr__(self):
-        return '<Device %s>' % self.name
+
