+++ api/run.py
@@ -7,7 +7,7 @@ def create_app(config_filename):
      from app import api_bp
      app.register_blueprint(api_bp, url_prefix='/api')
 
-     from Model import db
+     from models.device import db
      db.init_app(app)
 
      return app
